+++ contracts/mocks/GovernorCompatibilityBravoMock.sol
@@ -23,7 +23,7 @@ contract GovernorCompatibilityBravoMock is GovernorCompatibilityBravo, GovernorT
         return super.supportsInterface(interfaceId);
     }
 
-    function votingPeriod() public pure override(IGovernor, Governor) returns (uint64) {
+    function votingPeriod() public pure override(IGovernor, Governor) returns (uint256) {
         return 16; // blocks
     }
