+++ contracts/mocks/GovernorMock.sol
@@ -15,7 +15,7 @@ contract GovernorMock is Governor, GovernorWeightQuorumFractional, GovernorVotin
 
     receive() external payable {}
 
-    function votingPeriod() public pure override returns (uint64) {
+    function votingPeriod() public pure override returns (uint256) {
         return 16; // blocks
     }
